refactor(canvas): A*-Router-Effect aus CanvasArea in Hook auslagern (#468)
CanvasArea.tsx 2103 → 2016 Zeilen. Die ~85-zeilige A*-Router-Registrierung
(layoutOf/routeOne/routeAll + setCableRouter) wandert nach
useCanvasCableRouter.ts — eine kohärente, in sich geschlossene Concern
(nur Store-Instance + updateCable + mode + routerId).

Verhaltensneutral, beweisbar via Characterization-Test (Golden vorher auf
main aufgenommen, danach gegenübergestellt):
- Behavior-Signatur (Node-Positionen/-Größen, Edge-Pfade, Handle-Count,
  Selektion, Console-Errors) golden == after: IDENTISCH.
- Router-Funktion separat im Rack-Wire-Overlay verifiziert: Kabel 0→1,
  A*-Pfad mit Waypoints, 0 Page-Errors.
- tsc 0, eslint 0, build 0.

Refs #468
